Update files.rs
The u64::from() calls are the correct cross-platform widening, and the #[allow] attribute suppresses the clippy complaints on platforms where the fields are already u64

+++ src/utils/files.rs
@@ -210,7 +210,8 @@ fn check_disk_space(dir: &Path, needed_bytes: usize) -> Result<()> {
             let mut stat: libc::statvfs = std::mem::zeroed();
             if libc::statvfs(path_cstr.as_ptr(), &raw mut stat) == 0 {
                 #[allow(clippy::unnecessary_cast)]
-                let free_bytes = u64::from(stat.f_bavail) * stat.f_frsize;
+                #[allow(clippy::useless_conversion, clippy::cast_lossless)]
+                let free_bytes = u64::from(stat.f_bavail) * u64::from(stat.f_frsize);
                 let needed = (needed_bytes as u64).saturating_mul(2);
                 if free_bytes < needed {
                     return Err(anyhow::anyhow!(
